backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
alembic==1.14.0
reportlab==4.4.4
openpyxl==3.1.5
aiofiles==24.1.0
pytest==8.4.2
httpx==0.28.1
email-validator==2.2.0
